refactor(boards): mueve modules/board a features/boards; actions/board -> actions/reminders
