--- src/actions/board/index.ts
@@ -1,4 +1,0 @@
-export * from './getBoardById'
-export * from './updateBoardName'
-export * from './getReminders'
-export * from './saveReminders'
