+++ src/features/boards/api/repository/nextjsBoardRepository.ts
@@ -1,14 +1,14 @@
-import { boardModel } from '@/modules/board/models/board'
-import { BoardRepository } from '@/modules/board/repository/boardRepository'
+import { boardModel } from '../../model/board'
+import { BoardRepository } from './boardRepository'
 
 export default class NextjsBoardRepository implements BoardRepository {
 	async save(board: boardModel, boardId: string): Promise<void> {
-		const { updateBoardName } = await import('@/actions/board')
+		const { updateBoardName } = await import('../actions/updateBoardName')
 		await updateBoardName({ boardId, name: board.name })
 	}
 
 	async get(boardId: string): Promise<boardModel> {
-		const { getBoardById } = await import('@/actions/board')
+		const { getBoardById } = await import('../actions/getBoardById')
 		const board = await getBoardById({ boardId })
 		if (!board) throw new Error(`Board ${boardId} no encontrado`)
 		return board
