+++ src/features/boards/index.ts
@@ -1,0 +1,9 @@
+export type { boardModel } from './model/board'
+export {
+	defaultBoard,
+	isDefaultBoardName,
+	isThisBoardNameWithinTheLimitOfLetters,
+} from './model/board'
+export { Board } from './ui/Board'
+export { ChangeBoardName } from './ui/ChangeBoardName'
+export { useBoardQuery } from './hooks/useBoardQuery'
