+++ src/features/boards/ui/ChangeBoardName.tsx
@@ -5,11 +5,11 @@ import { ChangeEvent, KeyboardEvent, useState } from 'react'
 import { Button } from '@/shared/ui/atoms/button'
 import { PencilIcon } from '@/shared/ui/atoms/icons'
 import { Label } from '@/shared/ui/atoms/label'
-import { isThisBoardNameWithinTheLimitOfLetters } from '@/modules/board/models/board'
-import { changeBoardName } from '@/modules/board/useCase/changeBoardName'
+import { isThisBoardNameWithinTheLimitOfLetters } from '../model/board'
+import { changeBoardName } from '../useCase/changeBoardName'
 import getErrorMessageForTheUser from '@/shared/lib/getErrorMessageForTheUser'
 import { toast } from 'sonner'
-import { useBoardQuery } from '@/modules/board/hooks/useBoardQuery'
+import { useBoardQuery } from '../hooks/useBoardQuery'
 import { useTranslation } from 'react-i18next'
 import { SettingSection } from '@/shared/ui/organisms/SettingSection'
 import { useTheme } from '@/shared/hooks/useTheme'
