+++ src/views/BoardPage.tsx
@@ -1,13 +1,13 @@
 'use client'
 
-import { Board } from '@/modules/board/Board'
+import { Board } from '@/features/boards'
 import { ColumnsFooterContentProvider } from '@/modules/TaskBoard/components/Columns/context/ColumnsFooter/ColumnsFooterContentProvider'
 import { ColumnsFooterContent } from '@/modules/TaskBoard/components/Columns/context/ColumnsFooter/columnsFooterContent'
 import { ArchiveTaskListButton } from '@/modules/TaskBoard/components/taskList/components/ArchivedTasks/components/ArchiveTaskListButton'
 import { AddNewTaskInput } from '@/modules/TaskBoard/components/taskList/components/AddNewTaskInput'
 import { TaskListInEachColumn } from '@/modules/TaskBoard/components/taskList/TaskListInEachColumn'
 import PageContainer from './PageContainer'
-import { useBoardQuery } from '@/modules/board/hooks/useBoardQuery'
+import { useBoardQuery } from '@/features/boards'
 import { USER_IS_IN } from '@/shared/ui/organisms/userIsIn'
 import { ListView } from '@/modules/TaskBoard/components/ListView'
 import { TableView } from '@/modules/TaskBoard/components/TableView'
