+++ src/views/Settings.tsx
@@ -2,7 +2,7 @@
 
 import { useTranslation } from 'react-i18next'
 import { USER_IS_IN } from '../shared/ui/organisms/userIsIn'
-import { ChangeBoardName } from '../modules/board/components/ChangeBoardName'
+import { ChangeBoardName } from '@/features/boards'
 import CreateReminder from '@/modules/TaskBoard/components/Reminder/CreateReminder'
 import { ThemeSelection } from '@/shared/preferences/theme'
 import { ReminderList } from '@/modules/TaskBoard/components/Reminder/ReminderList'
@@ -12,7 +12,7 @@ import { EnableTags } from '@/features/tags'
 
 import { ConfigColumns } from '@/modules/TaskBoard/components/Columns/components/ConfigColumns'
 import { useTaskBoardQuery } from '@/modules/TaskBoard/hooks/useTaskBoardQuery'
-import { useBoardQuery } from '@/modules/board/hooks/useBoardQuery'
+import { useBoardQuery } from '@/features/boards'
 
 const useColumnListForReminders = (): { name: string; id: string; position: string }[] => {
 	const { taskBoard } = useTaskBoardQuery()
